sys-utility/requirements.txt
pyyaml
requests
psutil
pywin32
